Add vndb_id column in tags model and update my_rating type in novel schema
- Created a new migration script to add the `vndb_id` column in the tags model.
- Changed the type of `my_rating` in the NovelCreate schema from int to float for better precision.
- Updated test cases to reflect the change in `my_rating` type from int to float.

+++ backend/app/schemas/novel.py
@@ -43,6 +43,6 @@ class NovelsDetailResponse(BaseModel):
 class NovelCreate(BaseModel):
     status: str = "READING"
     my_review: Optional[str] = None
-    my_rating: Optional[int] = None
+    my_rating: Optional[float] = None
     language: str = "RUSSIAN"
     completed_date: Optional[date] = None
